Introduced MandlebrotCalculator class
- The purpose of this class is to manage the jobs.
- Created MandlebrotCalculator.h header file where Mandlebrot class is
declared as the child of QObject class.
- Introduced public and private variables, singnals and slots, to create
and destroy jobs and user their results to generate the picture. Refer
page 299-300 for details of each class member.
- Completed MandelbrotCalculator.cpp file that contains method to
properly create Job for each row of displayed area. Also create method
to abort the current Jobs, as well as send the resultant Job results to
GUI.